get rid of nestedsamplers completely
